converted negation to class to allow language specific implementation
