+++ src/test/scala/org/clulab/wm/eidos/system/TestEidosActions.scala
@@ -118,10 +118,6 @@ class TestEidosActions extends Test {
       (m.arguments(EidosActions.ANAPHOR).head.text == anaphor)
   }
 
-  def hasHedging(m: Mention, term: String): Boolean= {
-    val hedged = m.attachments.filter(_.isInstanceOf[Hedging])
-    hedged.exists(h => h.asInstanceOf[Hedging].trigger == term)
-  }
 
   def hasProperty(m: Mention, term: String): Boolean = {
     val props = m.attachments.filter(_.isInstanceOf[Property])
@@ -308,12 +304,6 @@ class TestEidosActions extends Test {
       val text = "Rainfall causes flooding.  This causes problems."
       val mentions = extractMentions(text)
       mentions.exists(m => correctCoreference(m, "flooding", "This"))
-    }
-
-    it should "identify hedging" in {
-      val text = "Rainfall is likely to cause flooding."
-      val mentions = extractMentions(text)
-      mentions.exists(m => hasHedging(m, "likely")) should be (true)
     }
 
     it should "identify properties" in {
