+++ src/test/scala/org/clulab/wm/eidos/system/TestHedging.scala
@@ -1,5 +1,21 @@
 package org.clulab.wm.eidos.system
 
-class TestHedging {
+import org.clulab.odin.Mention
+import org.clulab.wm.eidos.attachments.Hedging
+import org.clulab.wm.eidos.test.TestUtils._
+
+
+class TestHedging extends Test {
+
+  def hasHedging(m: Mention, term: String): Boolean= {
+    val hedged = m.attachments.filter(_.isInstanceOf[Hedging])
+    hedged.exists(h => h.asInstanceOf[Hedging].trigger == term)
+  }
+
+  it should "identify hedging" in {
+    val text = "Rainfall is likely to cause flooding."
+    val mentions = extractMentions(text)
+    mentions.exists(m => hasHedging(m, "likely")) should be (true)
+  }
 
 }
